\name{irc-package}
\alias{irc-package}
\alias{irc}
\docType{package}
\title{
  IRC
}
\description{
  IRC client protocol in R
}
\details{
\tabular{ll}{
Package: \tab irc\cr
Type: \tab Package\cr
Version: \tab 0.0.0\cr
Date: \tab 2014-12-09\cr
License: \tab BSD 3-clause \cr
}
}
\author{
Austin G. Davis-Richardson

Maintainer: Austin G. Davis-Richardson <harekrishna@gmail.com>
}
\keyword{ package irc bot chat }
\examples{
connect()
}
